翻译导入忽略中文为空的行

+++ script/translate_help_tool.py
@@ -97,22 +97,25 @@ def localize_import(root_path: str, mode_name: str, strings_name: str):
         if lang is None or len(lang) == 0:
             continue
         string_file_path = '{0}/{1}.lproj/{2}.strings'.format(root_path, lang, strings_name)
-        string_file = open(string_file_path, 'wt')
-        for j in range(2, max_row+1):
-            key = ws.cell(row=j, column=1).value
-            value = ws.cell(row=j, column=i).value
-            if value is None:
-                value = ""
-            if value != "":
-                key_placeholder_list = re.findall(format_string_placeholder_rex, key, re.RegexFlag.DOTALL)
-                if len(key_placeholder_list) != 0:
-                    value_placeholder_list = re.findall(format_string_placeholder_rex, value, re.RegexFlag.DOTALL)
-                    if key_placeholder_list != value_placeholder_list:
-                        print("占位符不匹配 模块: {0} 位置: 第{1}行 语言: {2}".format(mode_name, j, lang))
-                        value = ""
-            temp = '"{0}" = "{1}";\n'.format(key, value)
-            string_file.write(temp)
-        string_file.close()
+        if os.path.exists(string_file_path):
+            string_file = open(string_file_path, 'wt')
+            for j in range(2, max_row + 1):
+                key = ws.cell(row=j, column=1).value
+                if key is None:
+                    continue
+                value = ws.cell(row=j, column=i).value
+                if value is None:
+                    value = ""
+                if value != "":
+                    key_placeholder_list = re.findall(format_string_placeholder_rex, key, re.RegexFlag.DOTALL)
+                    if len(key_placeholder_list) != 0:
+                        value_placeholder_list = re.findall(format_string_placeholder_rex, value, re.RegexFlag.DOTALL)
+                        if key_placeholder_list != value_placeholder_list:
+                            print("占位符不匹配 模块: {0} 位置: 第{1}行 语言: {2}".format(mode_name, j, lang))
+                            value = ""
+                temp = '"{0}" = "{1}";\n'.format(key, value)
+                string_file.write(temp)
+            string_file.close()
 
 
 def parse_argv():
